requirements.txt
pyyaml>=6.0
boto3>=1.28
